+++ crates/zng-app/src/lib.rs
@@ -1189,6 +1189,8 @@ impl APP {
         assert_not_view_process();
         Self::assert_can_run();
         check_deadlock();
+        zng_env::init_process_name("app-process");
+
         let _ = INSTANT.now();
         let scope = LocalContext::start_app(AppId::new_unique());
         AppExtended {